Moved scheduler_base implementation to source file
- flyby: remove obsolete periodic_maintenance
