+++ examples/cancelable_action/cancelable_action/server/cancelable_action.hpp
@@ -10,6 +10,7 @@
 #include <hpx/include/components.hpp>
 #include <hpx/include/actions.hpp>
 #include <hpx/include/threads.hpp>
+#include <hpx/include/util.hpp>
 
 #include <atomic>
