--- hpx/runtime/threads/detail/periodic_maintenance.hpp
@@ -1,96 +1,0 @@
-//  Copyright (c) 2007-2015 Hartmut Kaiser
-//
-//  Distributed under the Boost Software License, Version 1.0. (See accompanying
-//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-
-#if !defined(HPX_RUNTIME_THREADS_DETAIL_PERIODIC_MAINTENANCE_JAN_11_2015_0626PM)
-#define HPX_RUNTIME_THREADS_DETAIL_PERIODIC_MAINTENANCE_JAN_11_2015_0626PM
-
-#include <hpx/config.hpp>
-#include <hpx/config/asio.hpp>
-#include <hpx/runtime_fwd.hpp>
-#include <hpx/state.hpp>
-#include <hpx/util/bind.hpp>
-#include <hpx/util/io_service_pool.hpp>
-#include <hpx/util/steady_clock.hpp>
-
-#include <boost/asio/basic_waitable_timer.hpp>
-
-#include <atomic>
-#include <chrono>
-#include <functional>
-#include <type_traits>
-
-namespace hpx { namespace threads { namespace detail
-{
-    ///////////////////////////////////////////////////////////////////////////
-    inline bool is_running_state(hpx::state state)
-    {
-        return state == state_running || state == state_suspended;
-    }
-
-    ///////////////////////////////////////////////////////////////////////////
-    template <typename SchedulingPolicy>
-    inline void periodic_maintenance_handler(SchedulingPolicy& scheduler,
-        std::atomic<hpx::state>& global_state, std::false_type)
-    {
-    }
-
-    template <typename SchedulingPolicy>
-    inline void periodic_maintenance_handler(SchedulingPolicy& scheduler,
-        std::atomic<hpx::state>& global_state, std::true_type)
-    {
-        bool running = is_running_state(global_state.load());
-        scheduler.periodic_maintenance(running);
-
-        if (running)
-        {
-            // create timer firing in correspondence with given time
-            typedef boost::asio::basic_waitable_timer<
-                util::steady_clock> deadline_timer;
-
-            deadline_timer t(
-                get_thread_pool("timer-thread")->get_io_service(),
-                std::chrono::milliseconds(1000));
-
-            void (*handler)(SchedulingPolicy&, std::atomic<hpx::state>&,
-                std::true_type) =
-                &periodic_maintenance_handler<SchedulingPolicy>;
-
-            t.async_wait(util::bind(handler, std::ref(scheduler),
-                std::ref(global_state), std::true_type()));
-        }
-    }
-
-    template <typename SchedulingPolicy>
-    inline void start_periodic_maintenance(SchedulingPolicy&,
-        std::atomic<hpx::state>& global_state, std::false_type)
-    {
-    }
-
-    template <typename SchedulingPolicy>
-    inline void start_periodic_maintenance(SchedulingPolicy& scheduler,
-        std::atomic<hpx::state>& global_state, std::true_type)
-    {
-        scheduler.periodic_maintenance(is_running_state(global_state.load()));
-
-        // create timer firing in correspondence with given time
-        typedef boost::asio::basic_waitable_timer<
-            util::steady_clock> deadline_timer;
-
-        deadline_timer t (
-            get_thread_pool("io-thread")->get_io_service(),
-            std::chrono::milliseconds(1000));
-
-        void (*handler)(SchedulingPolicy&, std::atomic<hpx::state>&,
-            std::true_type) =
-            &periodic_maintenance_handler<SchedulingPolicy>;
-
-        t.async_wait(util::bind(handler, std::ref(scheduler),
-            std::ref(global_state), std::true_type()));
-    }
-}}}
-
-#endif
-
-
